+++ defaults.py
@@ -125,8 +125,13 @@ class AppDefaults:
         "global_alt_sel_line": '#006E20BF',
         "global_draw_color": '#FF000080',
         "global_sel_draw_color": '#0000FF80',
-        "global_proj_item_color": '#000000FF',
-        "global_proj_item_dis_color": '#b7b7cbFF',
+
+        # Project Items colors
+        "global_proj_item_color_light": '#000000FF',
+        "global_proj_item_dis_color_light": '#b7b7cbFF',
+        "global_proj_item_color_dark": '#4385C8FF',
+        "global_proj_item_dis_color_dark": '#61616CFF',
+
         "global_project_autohide": True,
 
         # General App Settings